Refactor auth related notifications
